lista de usuarios se actualiza al agregar usuario
